- Transfer methods from report_compassion to child_compassion to complete the transfer of functionality. - Review patch
